Add function to drive individual bits of ss port
If the slave select line is wired to a multi-bit port the
drive_1bit_of_ss_port() function can be used to drive individual pins of
the ss port to desired values, while preserving the values on the other
pins.

Available in both the synchronous and asynchronous SPI master
interfaces.