Update slave to ISR based CS
States were found where cpha == 1 that would cause the thread to
deadlock on a port_out call.  This would happen in the condition where
the slave was provided more data to output, than the master was going to
clock.  To resolve this issue, CS is not ISR based.  The slave code that
reads in and drive out will still get stuck on a port_out call; however,
when the CS is deasserted the miso buffer is emptied, allowing the
port_out to finish on exit of the ISR.